+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * [Kernel] Ensure we don't splice keyword args unecessarily
 
 * deprecations
+  * [ExUnit] `assert left inlist right` is deprecated in favor of `assert left in right`
 
 * backwards incompatible changes
 
@@ -146,7 +147,6 @@
 
 * deprecations
   * [Enum] `Enum.qsort` is deprecated in favor of `Enum.sort`
-  * [ExUnit] `assert left in right` is deprecated in favor of `assert left inlist right`
   * [List] `List.sort` and `List.uniq` have been deprecated in favor of their `Enum` counterparts
   * [Record] Default-based generated functions are deprecated
   * [Typespec] Enhancements and deprecations to the `@spec/@callback` and the fun type syntax